Add basic Swift session support
A new keyword, `:session`, names a new or existing Swift REPL session
so that different code blocks can use the same execution environment.

This initial implementation filters some common output patterns from a
Swift REPL so that the result is cleaner.

TODO: We'd like to make this work with the `:toolchain` keyword.